+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 - Extended the survivalist achievement with 4 new stages
 - Added 1 new achievement!
+- Added a setting to adjust hotspot highlight border opacity
 
 ### Fixes
 - Fixed other messages being counted as trophy catches when they shouldn't